CREXDATA System Components/Text Analytics/Text Analytics Component/Sythetic-Data-Generation/requirements.txt
transformers
torch
pandas
huggingface-hub

--extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu 
llama-cpp-python